Criação da tela de pedido

+++ src/main/java/br/com/fdp/pedidos/model/ItemPedido.java
@@ -32,7 +32,7 @@ public class ItemPedido {
 	@ManyToOne
 	@JoinColumn(nullable=false)
 	private Package embrulho;
-	private long quantidade;
+	private short quantidade;
 	@Column(nullable = false, precision = 7, scale = 2)
 	private BigDecimal valorProduto;
 	@Column(nullable = false, precision = 7, scale = 2)
